Add payment channel update payment strategy
