+++ sdk/src/main/java/io/singularitynet/sdk/client/UpdatePaymentChannelPaymentStrategy.java
@@ -1,0 +1,18 @@
+package io.singularitynet.sdk.client;
+
+import lombok.ToString;
+
+import io.singularitynet.sdk.mpe.PaymentChannel;
+
+@ToString
+public class UpdatePaymentChannelPaymentStrategy extends PaymentChannelPaymentStrategy {
+        
+    public UpdatePaymentChannelPaymentStrategy() {
+    }
+
+    @Override
+    protected PaymentChannel selectChannel(ServiceClient serviceClient) {
+        throw new UnsupportedOperationException("Not implemented");
+    }
+
+}
